requirements.txt
opencv-contrib-python=4.4.0.44
numpy=1.19.2
pillow=7.2.0
pytesseract=0.3.6
